+++ src/CustomizeBuildings/Changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.0.60.0]
+- fixed empty settings when switching language; will now save settings separately for each language
+- fixed building substitution strings if they had no hyperlink
+- changed that pressure settings are also applied to "Supply Teleporter" (WarpConduitSender)
+
+## [1.0.59.0]
+- updated language file; thanks to Ventulus-lab
+
 ## [1.0.58.0] U39-490405-S
 - added rocket height tuning
 - fixed some tuning settings not being applied